@fix Make sure to destroy the loop scope when using control directives
This was causing a *very* peculiar bug that somehow only Linux actually choked on (it was present on other platforms too, but they just silently accepted it).
Basically, this was the problem: to keep track of objects and easily destroy them, we push them onto a global stack. To do that, we take their address (usually on the stack). So, the way the system's stack works is that variables occupy their space on the stack, and when they go out of scope, that space is availble again for the next variable to use. So, we weren't removing the loop variables from the stack when the loop went out of scope, and so that pointer on the stack was still pointing to that address on the stack. This has happened before, and in those cases, the next variable is a non-object variable and trying to unwind Alta's object stack has led to very straightforward errors. Unfortunately, in this particular instance, the next variable that occupied the loop variable's stack space was another object variable. Even better, since all objects have a common information structure, when the loop's variable finally got "unwound", it was actually calling the new variable's destructor, leading to later code trying to read from a free'd memory block from the object.
This would've been impossible to detect if not for 1) Ubuntu/Linux properly throwing a hissy fit over free'd memory access, and 2) Valgrind. Thanks Valgrind!
By the way, the offending transpiled code was that of `floatingPointToString` from the standard library's `util` package. The loop variable in question was the `char` variable and the variable being accidentally destroyed was the function's copied return value.