Update AltaCore and Ceetah...
The update to AltaCore includes a fix for ranged for loops not properly unwinding their stack (they were leaving the counter and the start and end tests intact, when these should be destroyed when the loop ends). This was having some seemingly unrelated side-effects, like for example: since the stack wasn't being unwound for these, they remained on the global runtime, and would be 'destroyed' at the end of their parent's loop. This meant that what was actually being destroyed were any variables on the stack in the parent scope that ended up with the same address. Yeah, it was pretty wack.