A couple of changes; please check the description
First, this commit implements the new enumeration lookup feature in AltaCore. All enumerations can do forward lookups, but only some can do reverse lookups (AltaCore already has us covered in this case; we won't be asked to do reverse lookups if the enumeration's underlying type doesn't support it). As such, to allow enumerations to be created even for types that can't be compared for equality, the reverse lookup function is only generated when they are able to do so (instead of throwing an error; like I said before, AltaCore already checks for this, so we won't need to do reverse lookups if the enumeration doesn't support it).

Second, this commit fixes a bug introduced in the previous commit which caused a double free error with temporary variables on the stack. It was a result of read accessors now tempifying their `self` arguments, and function calls automatically pushing their results onto the stack when not being assigned to a variable. The solution was to simply check if the `self` argument is a function call and, if so, then don't tempify it.

Third, this commit allows constant types to become optional, by removing the constant modifier on the `target` member of the optional structure.