Rework how generators store stack variables
They way we were doing it before was absolutely *bonkers*. The new way is basically to imitate what is done for regular functions: we allocate stacks for each scope, since we already know that most of the time we're going to use all the variables in that scope (the only time we might not is in loops). It's sufficiently memory efficient and it's much simpler and more stable than what we had before.